refactor: reuse dataframe code (part1)
