--- src/main/scala/org/globalforestwatch/summarystats/annualupdate_minimal/dataframes/IsoDownloadDF.scala
@@ -1,150 +1,0 @@
-package org.globalforestwatch.summarystats.annualupdate_minimal.dataframes
-
-import org.apache.spark.sql.functions.{round, sum}
-import org.apache.spark.sql.{DataFrame, SparkSession}
-
-object IsoDownloadDF {
-
-  def sumArea(df: DataFrame): DataFrame = {
-
-    val spark: SparkSession = df.sparkSession
-    import spark.implicits._
-
-    df.groupBy($"iso", $"treecover_density__threshold")
-      .agg(
-        sum($"treecover_extent_2000__ha") as "treecover_extent_2000__ha",
-        sum($"treecover_extent_2010__ha") as "treecover_extent_2010__ha",
-        sum($"area__ha") as "area__ha",
-        sum($"treecover_gain_2000-2012__ha") as "treecover_gain_2000-2012__ha",
-        sum($"aboveground_biomass_stock_2000__Mg") as "aboveground_biomass_stock_2000__Mg",
-        sum($"aboveground_biomass_stock_2000__Mg") / sum($"treecover_extent_2000__ha") as "avg_aboveground_biomass_2000_Mt_ha-1",
-        sum($"aboveground_co2_stock_2000__Mg") as "aboveground_co2_stock_2000__Mg",
-        sum($"treecover_loss_2001__ha") as "treecover_loss_2001__ha",
-        sum($"treecover_loss_2002__ha") as "treecover_loss_2002__ha",
-        sum($"treecover_loss_2003__ha") as "treecover_loss_2003__ha",
-        sum($"treecover_loss_2004__ha") as "treecover_loss_2004__ha",
-        sum($"treecover_loss_2005__ha") as "treecover_loss_2005__ha",
-        sum($"treecover_loss_2006__ha") as "treecover_loss_2006__ha",
-        sum($"treecover_loss_2007__ha") as "treecover_loss_2007__ha",
-        sum($"treecover_loss_2008__ha") as "treecover_loss_2008__ha",
-        sum($"treecover_loss_2009__ha") as "treecover_loss_2009__ha",
-        sum($"treecover_loss_2010__ha") as "treecover_loss_2010__ha",
-        sum($"treecover_loss_2011__ha") as "treecover_loss_2011__ha",
-        sum($"treecover_loss_2012__ha") as "treecover_loss_2012__ha",
-        sum($"treecover_loss_2013__ha") as "treecover_loss_2013__ha",
-        sum($"treecover_loss_2014__ha") as "treecover_loss_2014__ha",
-        sum($"treecover_loss_2015__ha") as "treecover_loss_2015__ha",
-        sum($"treecover_loss_2016__ha") as "treecover_loss_2016__ha",
-        sum($"treecover_loss_2017__ha") as "treecover_loss_2017__ha",
-        sum($"treecover_loss_2018__ha") as "treecover_loss_2018__ha",
-        sum($"aboveground_biomass_loss_2001__Mg") as "aboveground_biomass_loss_2001__Mg",
-        sum($"aboveground_biomass_loss_2002__Mg") as "aboveground_biomass_loss_2002__Mg",
-        sum($"aboveground_biomass_loss_2003__Mg") as "aboveground_biomass_loss_2003__Mg",
-        sum($"aboveground_biomass_loss_2004__Mg") as "aboveground_biomass_loss_2004__Mg",
-        sum($"aboveground_biomass_loss_2005__Mg") as "aboveground_biomass_loss_2005__Mg",
-        sum($"aboveground_biomass_loss_2006__Mg") as "aboveground_biomass_loss_2006__Mg",
-        sum($"aboveground_biomass_loss_2007__Mg") as "aboveground_biomass_loss_2007__Mg",
-        sum($"aboveground_biomass_loss_2008__Mg") as "aboveground_biomass_loss_2008__Mg",
-        sum($"aboveground_biomass_loss_2009__Mg") as "aboveground_biomass_loss_2009__Mg",
-        sum($"aboveground_biomass_loss_2010__Mg") as "aboveground_biomass_loss_2010__Mg",
-        sum($"aboveground_biomass_loss_2011__Mg") as "aboveground_biomass_loss_2011__Mg",
-        sum($"aboveground_biomass_loss_2012__Mg") as "aboveground_biomass_loss_2012__Mg",
-        sum($"aboveground_biomass_loss_2013__Mg") as "aboveground_biomass_loss_2013__Mg",
-        sum($"aboveground_biomass_loss_2014__Mg") as "aboveground_biomass_loss_2014__Mg",
-        sum($"aboveground_biomass_loss_2015__Mg") as "aboveground_biomass_loss_2015__Mg",
-        sum($"aboveground_biomass_loss_2016__Mg") as "aboveground_biomass_loss_2016__Mg",
-        sum($"aboveground_biomass_loss_2017__Mg") as "aboveground_biomass_loss_2017__Mg",
-        sum($"aboveground_biomass_loss_2018__Mg") as "aboveground_biomass_loss_2018__Mg",
-        sum($"aboveground_co2_emissions_2001__Mg") as "aboveground_co2_emissions_2001__Mg",
-        sum($"aboveground_co2_emissions_2002__Mg") as "aboveground_co2_emissions_2002__Mg",
-        sum($"aboveground_co2_emissions_2003__Mg") as "aboveground_co2_emissions_2003__Mg",
-        sum($"aboveground_co2_emissions_2004__Mg") as "aboveground_co2_emissions_2004__Mg",
-        sum($"aboveground_co2_emissions_2005__Mg") as "aboveground_co2_emissions_2005__Mg",
-        sum($"aboveground_co2_emissions_2006__Mg") as "aboveground_co2_emissions_2006__Mg",
-        sum($"aboveground_co2_emissions_2007__Mg") as "aboveground_co2_emissions_2007__Mg",
-        sum($"aboveground_co2_emissions_2008__Mg") as "aboveground_co2_emissions_2008__Mg",
-        sum($"aboveground_co2_emissions_2009__Mg") as "aboveground_co2_emissions_2009__Mg",
-        sum($"aboveground_co2_emissions_2010__Mg") as "aboveground_co2_emissions_2010__Mg",
-        sum($"aboveground_co2_emissions_2011__Mg") as "aboveground_co2_emissions_2011__Mg",
-        sum($"aboveground_co2_emissions_2012__Mg") as "aboveground_co2_emissions_2012__Mg",
-        sum($"aboveground_co2_emissions_2013__Mg") as "aboveground_co2_emissions_2013__Mg",
-        sum($"aboveground_co2_emissions_2014__Mg") as "aboveground_co2_emissions_2014__Mg",
-        sum($"aboveground_co2_emissions_2015__Mg") as "aboveground_co2_emissions_2015__Mg",
-        sum($"aboveground_co2_emissions_2016__Mg") as "aboveground_co2_emissions_2016__Mg",
-        sum($"aboveground_co2_emissions_2017__Mg") as "aboveground_co2_emissions_2017__Mg",
-        sum($"aboveground_co2_emissions_2018__Mg") as "aboveground_co2_emissions_2018__Mg"
-      )
-  }
-
-  def roundValues(df: DataFrame): DataFrame = {
-
-    val spark: SparkSession = df.sparkSession
-    import spark.implicits._
-
-    df.select(
-      $"iso" as "country",
-      $"treecover_density__threshold",
-      round($"treecover_extent_2000__ha") as "treecover_extent_2000__ha",
-      round($"treecover_extent_2010__ha") as "treecover_extent_2010__ha",
-      round($"area__ha") as "area__ha",
-      round($"treecover_gain_2000-2012__ha") as "treecover_gain_2000-2012__ha",
-      round($"aboveground_biomass_stock_2000__Mg") as "aboveground_biomass_stock_2000__Mg",
-      round($"avg_aboveground_biomass_2000_Mt_ha-1") as "avg_aboveground_biomass_2000_Mt_ha-1",
-      round($"aboveground_co2_stock_2000__Mg") as "aboveground_co2_stock_2000__Mg",
-      round($"treecover_loss_2001__ha") as "treecover_loss_2001__ha",
-      round($"treecover_loss_2002__ha") as "treecover_loss_2002__ha",
-      round($"treecover_loss_2003__ha") as "treecover_loss_2003__ha",
-      round($"treecover_loss_2004__ha") as "treecover_loss_2004__ha",
-      round($"treecover_loss_2005__ha") as "treecover_loss_2005__ha",
-      round($"treecover_loss_2006__ha") as "treecover_loss_2006__ha",
-      round($"treecover_loss_2007__ha") as "treecover_loss_2007__ha",
-      round($"treecover_loss_2008__ha") as "treecover_loss_2008__ha",
-      round($"treecover_loss_2009__ha") as "treecover_loss_2009__ha",
-      round($"treecover_loss_2010__ha") as "treecover_loss_2010__ha",
-      round($"treecover_loss_2011__ha") as "treecover_loss_2011__ha",
-      round($"treecover_loss_2012__ha") as "treecover_loss_2012__ha",
-      round($"treecover_loss_2013__ha") as "treecover_loss_2013__ha",
-      round($"treecover_loss_2014__ha") as "treecover_loss_2014__ha",
-      round($"treecover_loss_2015__ha") as "treecover_loss_2015__ha",
-      round($"treecover_loss_2016__ha") as "treecover_loss_2016__ha",
-      round($"treecover_loss_2017__ha") as "treecover_loss_2017__ha",
-      round($"treecover_loss_2018__ha") as "treecover_loss_2018__ha",
-      round($"aboveground_biomass_loss_2001__Mg") as "aboveground_biomass_loss_2001__Mg",
-      round($"aboveground_biomass_loss_2002__Mg") as "aboveground_biomass_loss_2002__Mg",
-      round($"aboveground_biomass_loss_2003__Mg") as "aboveground_biomass_loss_2003__Mg",
-      round($"aboveground_biomass_loss_2004__Mg") as "aboveground_biomass_loss_2004__Mg",
-      round($"aboveground_biomass_loss_2005__Mg") as "aboveground_biomass_loss_2005__Mg",
-      round($"aboveground_biomass_loss_2006__Mg") as "aboveground_biomass_loss_2006__Mg",
-      round($"aboveground_biomass_loss_2007__Mg") as "aboveground_biomass_loss_2007__Mg",
-      round($"aboveground_biomass_loss_2008__Mg") as "aboveground_biomass_loss_2008__Mg",
-      round($"aboveground_biomass_loss_2009__Mg") as "aboveground_biomass_loss_2009__Mg",
-      round($"aboveground_biomass_loss_2010__Mg") as "aboveground_biomass_loss_2010__Mg",
-      round($"aboveground_biomass_loss_2011__Mg") as "aboveground_biomass_loss_2011__Mg",
-      round($"aboveground_biomass_loss_2012__Mg") as "aboveground_biomass_loss_2012__Mg",
-      round($"aboveground_biomass_loss_2013__Mg") as "aboveground_biomass_loss_2013__Mg",
-      round($"aboveground_biomass_loss_2014__Mg") as "aboveground_biomass_loss_2014__Mg",
-      round($"aboveground_biomass_loss_2015__Mg") as "aboveground_biomass_loss_2015__Mg",
-      round($"aboveground_biomass_loss_2016__Mg") as "aboveground_biomass_loss_2016__Mg",
-      round($"aboveground_biomass_loss_2017__Mg") as "aboveground_biomass_loss_2017__Mg",
-      round($"aboveground_biomass_loss_2018__Mg") as "aboveground_biomass_loss_2018__Mg",
-      round($"aboveground_co2_emissions_2001__Mg") as "aboveground_co2_emissions_2001__Mg",
-      round($"aboveground_co2_emissions_2002__Mg") as "aboveground_co2_emissions_2002__Mg",
-      round($"aboveground_co2_emissions_2003__Mg") as "aboveground_co2_emissions_2003__Mg",
-      round($"aboveground_co2_emissions_2004__Mg") as "aboveground_co2_emissions_2004__Mg",
-      round($"aboveground_co2_emissions_2005__Mg") as "aboveground_co2_emissions_2005__Mg",
-      round($"aboveground_co2_emissions_2006__Mg") as "aboveground_co2_emissions_2006__Mg",
-      round($"aboveground_co2_emissions_2007__Mg") as "aboveground_co2_emissions_2007__Mg",
-      round($"aboveground_co2_emissions_2008__Mg") as "aboveground_co2_emissions_2008__Mg",
-      round($"aboveground_co2_emissions_2009__Mg") as "aboveground_co2_emissions_2009__Mg",
-      round($"aboveground_co2_emissions_2010__Mg") as "aboveground_co2_emissions_2010__Mg",
-      round($"aboveground_co2_emissions_2011__Mg") as "aboveground_co2_emissions_2011__Mg",
-      round($"aboveground_co2_emissions_2012__Mg") as "aboveground_co2_emissions_2012__Mg",
-      round($"aboveground_co2_emissions_2013__Mg") as "aboveground_co2_emissions_2013__Mg",
-      round($"aboveground_co2_emissions_2014__Mg") as "aboveground_co2_emissions_2014__Mg",
-      round($"aboveground_co2_emissions_2015__Mg") as "aboveground_co2_emissions_2015__Mg",
-      round($"aboveground_co2_emissions_2016__Mg") as "aboveground_co2_emissions_2016__Mg",
-      round($"aboveground_co2_emissions_2017__Mg") as "aboveground_co2_emissions_2017__Mg",
-      round($"aboveground_co2_emissions_2018__Mg") as "aboveground_co2_emissions_2018__Mg"
-    )
-  }
-}
